+++ lib/firebase_id_token/certificates.rb
@@ -1,22 +1,22 @@
 module FirebaseIdToken
   # Manage download and access of Google's x509 certificates. Keeps
-  # certificates on a Redis namespace database.
+  # certificates on an ActiveSupport cache.
   #
   # ## Download & Access Certificates
   #
   # It describes two ways to download it: {.request} and {.request!}.
-  # The first will only do something when Redis certificates database is empty,
+  # The first will only do something when the certificates cache is empty,
   # the second one will always request a new download to Google's API and
   # override the database with the response.
   #
   # It's important to note that when saving a set of certificates, it will also
-  # set a Redis expiration time to match Google's API header `expires`. **After
-  # this time went out, Redis will automatically delete those certificates.**
+  # set a expiration time to match Google's API header `expires`. **After
+  # this time went out, the cache will no longer provide those certificates.**
   #
   # *To know how many seconds left until the expiration you can use {.ttl}.*
   #
   # When comes to accessing it, you can either use {.present?} to check if
-  # there's any data inside Redis certificates database or {.all} to obtain an
+  # there's any data inside the cache or {.all} to obtain an
   # `Array` of current certificates.
   #
   # @example `.request` will only download once
@@ -30,16 +30,14 @@ module FirebaseIdToken
   #   FirebaseIdToken::Certificates.request! # Downloads certificates.
   #
   class Certificates
-    # A Redis instance.
-    attr_reader :redis
-    # Certificates saved in the Redis (JSON `String` or `nil`).
+    # Certificates saved in the cache (JSON `String` or `nil`).
     attr_reader :local_certs
 
     # Google's x509 certificates API URL.
     URL = 'https://www.googleapis.com/robot/v1/metadata/x509/'\
       'securetoken@system.gserviceaccount.com'
 
-    # Calls {.request!} only if there are no certificates on Redis. It will
+    # Calls {.request!} only if there are no certificates in the cache. It will
     # return `nil` otherwise.
     #
     # It will raise {Exceptions::CertificatesRequestError} if the request
@@ -49,11 +47,11 @@ module FirebaseIdToken
     # @return [nil, Hash]
     # @see Certificates.request!
     def self.request
-      new.request
+      new_child.request
     end
 
     # Triggers a HTTPS request to Google's x509 certificates API. If it
-    # responds with a status `200 OK`, saves the request body into Redis and
+    # responds with a status `200 OK`, saves the request body into the cache and
     # returns it as a `Hash`.
     #
     # Otherwise it will raise a {Exceptions::CertificatesRequestError}.
@@ -63,7 +61,7 @@ module FirebaseIdToken
     # {Exceptions::CertificatesTtlError}. You are mostly like to never face it.
     # @return [Hash]
     def self.request!
-      new.request!
+      new_child.request!
     end
 
     # @deprecated Use only `request!` in favor of Ruby conventions.
@@ -73,38 +71,38 @@ module FirebaseIdToken
       warn 'WARNING: FirebaseIdToken::Certificates.request_anyway is '\
         'deprecated. Use FirebaseIdToken::Certificates.request! instead.'
 
-      new.request!
+      new_child.request!
     end
 
-    # Returns `true` if there's certificates data on Redis, `false` otherwise.
+    # Returns `true` if there's certificates data in the cache, `false` otherwise.
     # @example
     #   FirebaseIdToken::Certificates.present? #=> false
     #   FirebaseIdToken::Certificates.request
     #   FirebaseIdToken::Certificates.present? #=> true
     def self.present?
-      ! new.local_certs.empty?
+      ! new_child.local_certs.empty?
     end
 
     # Returns an array of hashes, each hash is a single `{key => value}` pair
     # containing the certificate KID `String` as key and a
     # `OpenSSL::X509::Certificate` object of the respective certificate as
-    # value. Returns a empty `Array` when there's no certificates data on
-    # Redis.
+    # value. Returns a empty `Array` when there's no certificates data in
+    # the cache.
     # @return [Array]
     # @example
     #   FirebaseIdToken::Certificates.request
     #   certs = FirebaseIdToken::Certificates.all
     #   certs.first #=> {"1d6d01c7[...]" => #<OpenSSL::X509::Certificate[...]}
     def self.all
-      new.local_certs.map { |kid, cert|
+      new_child.local_certs.map { |kid, cert|
         { kid => OpenSSL::X509::Certificate.new(cert) } }
     end
 
     # Returns a `OpenSSL::X509::Certificate` object of the requested Key ID
     # (KID) if there's one. Returns `nil` otherwise.
     #
-    # It will raise a {Exceptions::NoCertificatesError} if the Redis
-    # certificates database is empty.
+    # It will raise a {Exceptions::NoCertificatesError} if the
+    # certificates cache is empty.
     # @param [String] kid Key ID
     # @return [nil, OpenSSL::X509::Certificate]
     # @example
@@ -112,7 +110,7 @@ module FirebaseIdToken
     #   cert = FirebaseIdToken::Certificates.find "1d6d01f4w7d54c7[...]"
     #   #=> <OpenSSL::X509::Certificate: subject=#<OpenSSL [...]
     def self.find(kid, raise_error: false)
-      certs = new.local_certs
+      certs = new_child.local_certs
       raise Exceptions::NoCertificatesError if certs.empty?
 
       return OpenSSL::X509::Certificate.new certs[kid] if certs[kid]
@@ -128,8 +126,8 @@ module FirebaseIdToken
     #
     # @raise {Exceptions::CertificateNotFound} if it cannot be found.
     #
-    # @raise {Exceptions::NoCertificatesError} if the Redis certificates
-    # database is empty.
+    # @raise {Exceptions::NoCertificatesError} if the certificates cache
+    # is empty.
     #
     # @param [String] kid Key ID
     # @return [OpenSSL::X509::Certificate]
@@ -146,17 +144,22 @@ module FirebaseIdToken
     # time, use it to know when to request again.
     # @return [Fixnum]
     def self.ttl
-      ttl = new.redis.ttl('certificates')
-      ttl < 0 ? 0 : ttl
+      # call a child class based on the configuration
+      klass = FirebaseIdToken.configuration.klass
+      klass.ttl
     end
 
-    # Sets two instance attributes: `:redis` and `:local_certs`. Those are
-    # respectively a Redis instance from {FirebaseIdToken::Configuration} and
+    def self.new_child
+      klass = FirebaseIdToken.configuration.klass
+      klass.new
+    end
+
+    # Sets two instance attributes: `:cach_store` and `:local_certs`. Those are
+    # respectively a cache instance from {FirebaseIdToken::Configuration} and
     # the certificates in it.
     def initialize
-      @redis = Redis::Namespace.new('firebase_id_token',
-        redis: FirebaseIdToken.configuration.redis)
-      @local_certs = read_certificates
+      # this should not be called directly. Call a child class
+      raise NotImplementedError
     end
 
     # @see Certificates.request
@@ -172,29 +175,6 @@ module FirebaseIdToken
         save_certificates
       else
         raise Exceptions::CertificatesRequestError.new(code)
-      end
-    end
-
-    private
-
-    def read_certificates
-      certs = @redis.get 'certificates'
-      certs ? JSON.parse(certs) : {}
-    end
-
-    def save_certificates
-      @redis.setex 'certificates', ttl, @request.body
-      @local_certs = read_certificates
-    end
-
-    def ttl
-      cache_control = @request.headers['cache-control']
-      ttl = cache_control.match(/max-age=([0-9]+)/).captures.first.to_i
-
-      if ttl > 3600
-        ttl
-      else
-        raise Exceptions::CertificatesTtlError
       end
     end
   end