+++ lib/firebase_id_token/exceptions/unsupported_cache_operation_error.rb
@@ -1,0 +1,11 @@
+module FirebaseIdToken
+  module Exceptions
+    # @see FirebaseIdToken::Certificates.request
+    # @see FirebaseIdToken::Certificates.request!
+    class UnsupportedCacheOperationError < StandardError
+      def initialize(message = "Cache store does not support a TTL read on entries")
+        super message
+      end
+    end
+  end
+end